+++ pygccxml/declarations/namespace.py
@@ -8,6 +8,8 @@ Describe a C++ namespace declaration.
 
 """
 
+import warnings
+
 from . import scopedef
 from . import declaration_utils
 
@@ -141,7 +143,19 @@ class namespace_t(scopedef.scopedef_t):
                 recursive=recursive,
                 allow_empty=allow_empty)
         )
-    nss = namespaces
+
+    def nss(self, name=None, function=None, recursive=None, allow_empty=None):
+        """
+        Deprecated method. Use the namespaces() method instead.
+
+        Deprecated since v1.9.0. Will be removed in v2.0.0
+        """
+        warnings.warn(
+            "The nss() method is deprecated. \n" +
+            "Please use the namespaces() method instead.",
+            DeprecationWarning)
+
+        return self.namespaces(name, function, recursive, allow_empty)
 
     def free_function(
             self,
@@ -170,7 +184,31 @@ class namespace_t(scopedef.scopedef_t):
                 header_file=header_file,
                 recursive=recursive)
         )
-    free_fun = free_function
+
+    def free_fun(
+            self,
+            name=None,
+            function=None,
+            return_type=None,
+            arg_types=None,
+            header_dir=None,
+            header_file=None,
+            recursive=None):
+        """
+        Deprecated method. Use the free_function() method instead.
+
+        Deprecated since v1.9.0. Will be removed in v2.0.0
+        """
+        warnings.warn(
+            "The free_fun() method is deprecated. \n" +
+            "Please use the free_function() method instead.",
+            DeprecationWarning)
+
+        return self.free_function(
+            name, function,
+            return_type, arg_types,
+            header_dir, header_file,
+            recursive)
 
     def free_functions(
             self,
@@ -201,7 +239,32 @@ class namespace_t(scopedef.scopedef_t):
                 recursive=recursive,
                 allow_empty=allow_empty)
         )
-    free_funs = free_functions
+
+    def free_funs(
+            self,
+            name=None,
+            function=None,
+            return_type=None,
+            arg_types=None,
+            header_dir=None,
+            header_file=None,
+            recursive=None,
+            allow_empty=None):
+        """
+        Deprecated method. Use the free_functions() method instead.
+
+        Deprecated since v1.9.0. Will be removed in v2.0.0
+        """
+        warnings.warn(
+            "The free_funs() method is deprecated. \n" +
+            "Please use the free_functions() method instead.",
+            DeprecationWarning)
+
+        return self.free_functions(
+            name, function,
+            return_type, arg_types,
+            header_dir, header_file,
+            recursive, allow_empty)
 
     def free_operator(
             self,