+++ unittests/const_volatile_arg_tester.py
@@ -27,7 +27,7 @@ class Test(parser_test_case.parser_test_case_t):
         self.global_ns = Test.global_ns
 
     def test(self):
-        f = self.global_ns.free_fun('pygccxml_bug')
+        f = self.global_ns.free_function('pygccxml_bug')
         t = f.arguments[0].decl_type
         self.assertTrue(isinstance(t, declarations.pointer_t))
         self.assertTrue(isinstance(t.base, declarations.volatile_t))