+++ unittests/patcher_tester.py
@@ -23,7 +23,7 @@ class tester_impl_t(parser_test_case.parser_test_case_t):
         self.__cxx_std = utils.cxx_standard(self.config.cflags)
 
     def test_enum_patcher(self):
-        fix_enum = self.global_ns.free_fun("fix_enum")
+        fix_enum = self.global_ns.free_function("fix_enum")
         default_val = fix_enum.arguments[0].default_value
         if self.__cxx_std.is_cxx11_or_greater:
             val = "::ns1::ns2::fruit::apple"
@@ -33,31 +33,31 @@ class tester_impl_t(parser_test_case.parser_test_case_t):
 
         if 32 == self.architecture or \
                 self.xml_generator_from_xml_file.is_castxml:
-            fix_enum2 = self.global_ns.free_fun("fix_enum2")
+            fix_enum2 = self.global_ns.free_function("fix_enum2")
             default_val = fix_enum2.arguments[0].default_value
             self.assertEqual(default_val, val)
 
             ns1 = self.global_ns.namespace("ns1")
             ns2 = ns1.namespace("ns2")
-            fix_enum2 = ns2.free_fun("fix_enum2")
+            fix_enum2 = ns2.free_function("fix_enum2")
             default_val = fix_enum2.arguments[0].default_value
             self.assertEqual(default_val, val)
 
-            fix_enum3 = self.global_ns.free_fun("fix_enum3")
+            fix_enum3 = self.global_ns.free_function("fix_enum3")
             default_val = fix_enum3.arguments[0].default_value
             val = val.replace("apple", "orange")
             self.assertEqual(default_val, val)
 
             if self.__cxx_std.is_cxx11_or_greater:
-                fix_enum4 = self.global_ns.free_fun("fix_enum4")
+                fix_enum4 = self.global_ns.free_function("fix_enum4")
                 default_val = fix_enum4.arguments[0].default_value
                 self.assertEqual(default_val, "::ns4::color::blue")
 
-                fix_enum5 = self.global_ns.free_fun("fix_enum5")
+                fix_enum5 = self.global_ns.free_function("fix_enum5")
                 default_val = fix_enum5.arguments[0].default_value
                 self.assertEqual(default_val, "::ns4::color::blue")
 
-            lpe = self.global_ns.free_fun("log_priority_enabled")
+            lpe = self.global_ns.free_function("log_priority_enabled")
             default_val = lpe.arguments[0].default_value
             if self.__cxx_std.is_cxx11_or_greater:
                 val = "(long int)" + \
@@ -127,7 +127,7 @@ class tester_impl_t(parser_test_case.parser_test_case_t):
                 fun1.arguments[1].default_value,
                 "ns1::st1::DEFAULT_2")
 
-        fun2 = self.global_ns.free_fun("fun2")
+        fun2 = self.global_ns.free_function("fun2")
         self.assertEqual(
             fun2.arguments[0].default_value,
             "::DEFAULT_1")
@@ -156,12 +156,12 @@ class tester_impl_t(parser_test_case.parser_test_case_t):
                 "ns1::st1::DEFAULT_2")
 
     def test_unnamed_enum_patcher(self):
-        fix_unnamed = self.global_ns.free_fun("fix_unnamed")
+        fix_unnamed = self.global_ns.free_function("fix_unnamed")
         self.assertEqual(
             fix_unnamed.arguments[0].default_value, "int(::fx::unnamed)")
 
     def test_function_call_patcher(self):
-        fix_function_call = self.global_ns.free_fun("fix_function_call")
+        fix_function_call = self.global_ns.free_function("fix_function_call")
         default_val = fix_function_call.arguments[0].default_value
         if self.xml_generator_from_xml_file.is_castxml:
             output_verion = self.xml_generator_from_xml_file.xml_output_version
@@ -178,7 +178,7 @@ class tester_impl_t(parser_test_case.parser_test_case_t):
         self.assertEqual(default_val, val)
 
     def test_fundamental_patcher(self):
-        fcall = self.global_ns.free_fun("fix_fundamental")
+        fcall = self.global_ns.free_function("fix_fundamental")
         if self.__cxx_std.is_cxx11_or_greater:
             val = "(unsigned int)(::fundamental::spam::eggs)"
         else:
@@ -187,7 +187,7 @@ class tester_impl_t(parser_test_case.parser_test_case_t):
             fcall.arguments[0].default_value, val)
 
     def test_constructor_patcher(self):
-        typedef__func = self.global_ns.free_fun("typedef__func")
+        typedef__func = self.global_ns.free_function("typedef__func")
         default_val = typedef__func.arguments[0].default_value
         if self.xml_generator_from_xml_file.is_gccxml_09 or \
                 self.xml_generator_from_xml_file.is_gccxml_09_buggy:
@@ -199,7 +199,7 @@ class tester_impl_t(parser_test_case.parser_test_case_t):
             val = "::typedef_::alias( )"
         self.assertEqual(default_val, val)
         if 32 == self.architecture:
-            clone_tree = self.global_ns.free_fun("clone_tree")
+            clone_tree = self.global_ns.free_function("clone_tree")
             if not self.xml_generator_from_xml_file.is_gccxml_09 and \
                     not self.xml_generator_from_xml_file.is_gccxml_09_buggy:
                 default_values = [